requirements.txt
asgiref==3.2.7
Django==2.2.1
ImageHash==4.0
numpy==1.16.3
Pillow==6.0.0
psycopg2==2.8.4
pytz==2019.1
PyWavelets==1.0.3
scipy==1.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.8
python-memcached==1.59